GradientRestApi/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-jwt==5.1.0